Refactor image filesystem processing
Separate image paths, tar reading, layer scanning, and entry models from filesystem orchestration. Decompose extraction into planning and ordered stages while preserving rollback and cancellation behavior.


Copilot-Session: fca5d6b0-eca9-4a5c-9474-eb8dbd0a7896
